+++ lib/neuron/include/neuron.hpp
@@ -6,16 +6,16 @@
 class neuron {
    private:
     matrix* weights;
-    float u;
+    double u;
 
    public:
     neuron(int numberOfInputs);
     /**
      * @brief multiplica los inputs por la matriz de pesos
      * @param input
-     * @return float result o NULL si no se puede calcular
+     * @return double result o NULL si no se puede calcular
      */
-    float calculate(matrix* input);
+    double calculate(matrix* input);
 };
 
 #endif
